+++ src/game/sprites/explosion.cc
@@ -28,22 +28,22 @@ const int       Explosion::HERO_FIREBALL_WEAPON = 0;
 const int       Explosion::HERO_EXPLOSION_WEAPON = 1;
 uint32          Explosion::WEAPON_ANIMATIONS[2];
 
-COLLISION_DIRECT(float, DamageCollision, obj) {
+COLLISION_DIRECT(double, DamageCollision, obj) {
 	Creature *creature = (Creature *) obj;
     creature->TakeDamage(data_);
 }
 
-Explosion::Explosion(ugdk::graphic::FlexibleSpritesheet *image, uint32 animation, float radius, float damage)
+Explosion::Explosion(ugdk::graphic::FlexibleSpritesheet *image, uint32 animation, double radius, double damage)
 {
 
     ugdk::graphic::Sprite* sprite = new ugdk::graphic::Sprite(image, ANIMATIONS);
     node_->set_drawable(sprite);
     image->set_hotspot(Vector2D(CENTER_X, CENTER_Y));
 	damage_ = damage;
-	bound_ = new pyramidworks::geometry::Circle(radius / 2.0f);
+	bound_ = new pyramidworks::geometry::Circle(radius / 2.0);
 	set_light_radius(1.3*radius);
 
-    Color light_color(1.0f, 0.521568f, 0.082352f);
+    Color light_color(1.0, 0.521568, 0.082352);
     node_->light()->set_color(light_color);
 
 	sprite->AddObserverToAnimation(this);
@@ -80,18 +80,18 @@ void Explosion::Tick() {
     this->status_ = WorldObject::STATUS_DEAD;
 }
 
-void Explosion::RadiusUpdate(float delta_t) {
+void Explosion::RadiusUpdate(double delta_t) {
 	bound_->set_radius(bound_->radius() + expansion_speed_ * delta_t);
 }
 
-void Explosion::Update(float delta_t) {
+void Explosion::Update(double delta_t) {
     WorldObject::Update(delta_t);
 	this->RadiusUpdate(delta_t);
 }
 
 void Explosion::set_world_position(const ugdk::Vector2D& pos) {
     super::set_world_position(pos);
-    node_->set_zindex(node_->zindex() + node_->drawable()->height() * 0.1f);
+    node_->set_zindex(node_->zindex() + node_->drawable()->height() * 0.1);
 }
 
 }
